refactor: verifier 패키지 변경, domain -> application
